refactor(bootstrap): 모듈 이름을 starter로 변경하고 패키지 위치 수정
