--- bootstrap/application-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/application/ApplicationProperties.java
@@ -1,19 +1,0 @@
-package com.seatliberator.seatliberator.bootstrap.application;
-
-import jakarta.validation.constraints.NotNull;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.bind.DefaultValue;
-import org.springframework.validation.annotation.Validated;
-
-import java.time.ZoneId;
-
-@Validated
-@ConfigurationProperties(prefix = "seatliberator.bootstrap.application")
-public record ApplicationProperties(
-        @DefaultValue("true")
-        boolean enabled,
-
-        @DefaultValue("UTC")
-        @NotNull ZoneId timezone
-) {
-}
