--- bootstrap/application-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-com.seatliberator.seatliberator.bootstrap.application.ApplicationAutoConfiguration
