--- bootstrap/security-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/security/customizer/HttpSecurityCustomizer.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.bootstrap.security.customizer;
-
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-
-public interface HttpSecurityCustomizer {
-    void customize(HttpSecurity httpSecurity) throws Exception;
-}
