+++ identity/identity-server/identity-server-application/build.gradle.kts
@@ -15,7 +15,7 @@ dependencies {
     implementation(project(":reservation:reservation-api"))
     implementation(project(":board:board-api"))
     implementation(project(":event-relay:event-relay-starter"))
-    implementation(project(":bootstrap:application-starter"))
+    implementation(project(":starter:application-starter"))
 
     implementation(libs.spring.boot.starter.oauth2.client)
     implementation(libs.spring.boot.starter.kafka)
