+++ starter/application-starter/build.gradle.kts
