+++ starter/application-starter/src/main/java/com/seatliberator/seatliberator/starter/application/ApplicationAutoConfiguration.java
@@ -1,0 +1,25 @@
+package com.seatliberator.seatliberator.starter.application;
+
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+
+import java.time.Clock;
+
+@AutoConfiguration
+@ConditionalOnProperty(
+        prefix = "seatliberator.bootstrap.application",
+        name = "enabled",
+        havingValue = "true",
+        matchIfMissing = true
+)
+@EnableConfigurationProperties(ApplicationProperties.class)
+public class ApplicationAutoConfiguration {
+    @Bean
+    @ConditionalOnMissingBean(Clock.class)
+    Clock clock(ApplicationProperties properties) {
+        return Clock.system(properties.timezone());
+    }
+}
