+++ starter/security-starter/src/main/java/com/seatliberator/seatliberator/starter/security/customizer/HttpSecurityCustomizer.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.starter.security.customizer;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+
+public interface HttpSecurityCustomizer {
+    void customize(HttpSecurity httpSecurity) throws Exception;
+}
